testing GPIO handling in an RTOS task.
Not sure how this will fit into the rest of the application just yet. I
may end up leaving the tasks separate from the C++ classes, similar to
this. It's also an option to wrap tasks in C++, but I'm not sure which
route is better yet. Ideally, this would be more integrated with the
RunLogger class.