simplify debouce code
Still not bullet-proof, but I like this better than the earlier
approach. Now, I just throw away all presses shorter than 1 second. Less
than this still has some false positives. I'd rather it throw away a
legitmate press than to potentially start and pause immediately and
possibly not record an entire run. So in other words, now, if you press
it more than once in a one-second period, it will only recognize the
first press. I like this better because I'm very unlikely to want to
record a 1 second run, and I want to be absolutely sure that it starts
recording when I press the button.