change sleep interval to 60min, and keep db connection alive.
The sleep interval is just so that the analytics script will be able to
update more frequently without much lag from when a new run event is
added. One hour seems like a decent tradeoff.

Removing the db connection.end() call fixes a bug where the warehouse
script crashes if it runs continuously and queries the database more
than once.